requirements.txt
fastapi
strawberry-graphql
sqlalchemy
asyncpg
pydantic
pyjwt
bcrypt
sentence-transformers
